refactoring withdraw command
