+++ Cli/Commands/TransactionCommands.cs
@@ -103,107 +103,6 @@ namespace Cli.Commands
         }
     }
 
-
-    public class WithdrawCommand : Command<WithdrawCommand.Settings>
-    {
-        public class Settings : CommandSettings
-        {
-
-        }
-
-        public override int Execute(CommandContext context, Settings settings)
-        {
-            var reference = CliWidgets.PromptText("Enter a reference for your transaction");
-            if (string.IsNullOrEmpty(reference))
-            {
-                CliWidgets.RenderError("[red]Reference must be specified.[/]");
-                return 1;
-            }
-
-            if (!int.TryParse(CliWidgets.PromptText("Transfer Amount"), out int amount))
-            {
-                CliWidgets.RenderError("Enter a valid integer");
-                return 1;
-            }
-            if (amount <= 0)
-            {
-                CliWidgets.RenderError("[red]Amount must be greater than zero.[/]");
-                return 1;
-            }
-            using var httpClient = new HttpClient();
-
-            // Add the Authorization header with the bearer token
-            var bearerToken = User.Token; // Replace with the actual token
-            httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", bearerToken);
-
-            try
-            {
-                // Fetch accounts from the API
-                // TODO: fetch api from configuration
-                // TODO: Stop using email endpoint
-                // TODO: api endpoints should start with /api
-                // CUSTOM: Use account number instead
-                var response = httpClient.GetAsync($"{Constants.ApiBaseUrl}/accounts").Result;
-
-                if (!response.IsSuccessStatusCode)
-                {
-                    CliWidgets.RenderError($"[red]Failed to fetch accounts: {response.StatusCode} - {response.ReasonPhrase}[/]");
-                    return 1;
-                }
-
-                var jsonResponse = response.Content.ReadAsStringAsync().Result;
-                var accounts = JsonSerializer.Deserialize<List<Account>>(jsonResponse);
-
-                if (accounts == null || !accounts.Any())
-                {
-                    CliWidgets.RenderWarning("[yellow]No accounts found.[/]");
-                    return 1;
-                }
-
-                // Prepare account choices
-                var accountChoices = accounts.Select(a => $"{a.AccountNumber} - {a.AccountType.Name}").ToList();
-
-                // Prompt user to select an account
-                var selectedAccount = CliWidgets.RenderSelection("Select an account", accountChoices);
-
-                // Extract AccountId from the selected choice
-                var accountNumber = selectedAccount.Split(" - ")[0];
-
-                var payload = new
-                {
-                    AccountNumber = accountNumber,
-                    Amount = amount,
-                    Reference = reference
-                };
-
-                var jsonPayload = JsonSerializer.Serialize(payload);
-                var content = new StringContent(jsonPayload, Encoding.UTF8, "application/json");
-
-                // Send the deposit/withdraw request
-                var endpoint = this.GetType().Name == nameof(DepositCommand) ? "deposits" : "withdraws";
-                var result = httpClient.PostAsync($"{Constants.ApiBaseUrl}/{endpoint}", content).Result;
-
-                if (result.IsSuccessStatusCode)
-                {
-                    CliWidgets.RenderPanel($"[green]{(endpoint == "deposit" ? "Deposited" : "Withdrawn")} Q {amount:n0} to account {accountNumber} with reference {reference}[/]");
-                    return 0;
-                }
-                else
-                {
-                    var errorMessage = result.Content.ReadAsStringAsync().Result;
-                    CliWidgets.RenderError($"[red]Failed to {endpoint}: {result.StatusCode} - {result.ReasonPhrase} - {errorMessage}[/]");
-                    return 1;
-                }
-            }
-            catch (Exception ex)
-            {
-                CliWidgets.RenderError($"[red]An error occurred: {ex.Message}[/]");
-                return 1;
-            }
-        }
-    }
-
-
     public class GetAllTransactionsCommand : Command<GetAllTransactionsCommand.Settings>
     {
         public class Settings : CommandSettings
